move periphGpioPin to its own file
